PDJB-1168: recombine cyas
