+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/landlordRegistration/states/LandlordRegistrationState.kt
@@ -1,8 +1,7 @@
 package uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.states
 
-import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.IndividualLandlordRegistrationCyaStep
+import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.LandlordRegistrationCyaStep
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.LandlordTypeStep
-import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.OrgLandlordRegistrationCyaStep
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.PrivacyNoticeStep
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks.IdentityTask
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks.IndividualLandlordRegistrationTask
@@ -18,6 +17,5 @@ interface LandlordRegistrationState :
     val privacyNoticeStep: PrivacyNoticeStep
     val identityTask: IdentityTask
     override val finishCyaStep: FinishCyaJourneyStep
-    override val cyaStep: IndividualLandlordRegistrationCyaStep
-    val orgCyaStep: OrgLandlordRegistrationCyaStep
+    override val cyaStep: LandlordRegistrationCyaStep
 }
